requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
jsonschema==3.2.0
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
pyrsistent==0.20.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.10.2
six==1.17.0
spectacular==0.4.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
